AI Model/civicsafe-ai/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
google-genai
pytest
python-multipart
